# src/swagger.yaml
openapi: 3.0.0
info:
  title: OKR Express API
  version: 1.0.0
  description: API for managing OKRs, companies, and teams
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        companies:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
        roles:
          type: array
          items:
            type: object
            properties:
              company:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
              role:
                type: string
    Team:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        companyId:
          type: string
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
        requiredRoles:
          type: array
          items:
            type: string
...
paths:
...
  /api/teams:
    post:
      summary: Create a new team
      tags: [Teams]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - companyId
              properties:
                name:
                  type: string
                companyId:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Company not found
        '400':
          description: Invalid input data

    get:
      summary: Get all teams
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: string
          description: ID of the company to get teams from
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '400':
          description: Company ID is required
        '404':
          description: Company not found
  /api/teams/{teamId}:
    get:
      summary: Get team details
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Team not found

    delete:
      summary: Delete team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
  /api/teams/{teamId}/members:
    get:
      summary: Get team members
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
...
  /api/teams/{teamId}/users/{userId}:
    delete:
      summary: Remove user from team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User removed from team successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
        '400':
          description: User is not in this team
...
  /api/teams/{teamId}/roles/{roleName}:
    delete:
      summary: Remove required role from team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
        - in: path
          name: roleName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
        '400':
          description: Role is not required for this team
...
  /api/teams/{teamId}/roles:
    post:
      summary: Add a single required role to team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roleName
              properties:
                roleName:
                  type: string
                  description: Name of the role to add
      responses:
        '201':
          description: Role added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: Role is already required for this team
        '401':
          description: Unauthorized
        '404':
          description: Team not found or Role not found in company
  /api/teams/{teamId}/users:
    post:
      summary: Add a single user to team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: ID of the user to add
      responses:
        '201':
          description: User added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: User is already in team or does not have required roles
        '401':
          description: Unauthorized
        '404':
          description: Team not found or User not found
  /api/teams/{teamId}/roles/bulk:
    post:
      summary: Add multiple required roles to team (bulk operation)
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roles
              properties:
                roles:
                  type: array
                  items:
                    type: string
                  description: Array of role names to add (use /roles endpoint for single role)
      responses:
        '201':
          description: Roles added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: Invalid roles or all roles already required
        '401':
          description: Unauthorized
        '404':
          description: Team not found or Company not found
  /api/teams/{teamId}/users/bulk:
    post:
      summary: Add multiple users to team (bulk operation)
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userIds
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                  description: Array of user IDs to add (use /users endpoint for single user)
      responses:
        '201':
          description: Users added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: Invalid users or all users already in team
        '401':
          description: Unauthorized
        '404':
          description: Team not found or some users not found 